requirements.txt
requests==2.25.1
APScheduler==3.7.0
beautifulsoup4==4.10.0
pycryptodome==3.10.1
